+++ utils/stribri.sh
@@ -1,2 +1,2 @@
 #!/bin/sh
-xterm -geometry 80x16 -title stribri -e "~/projecten/striso/software/striso_bridge/stribri_nox.sh" &
+xterm -geometry 80x16 -title stribri -e "./stribri_nox.sh" &