+++ utils/stribri_nox.sh
@@ -1,2 +1,2 @@
 #!/bin/sh
-~/projecten/striso/software/striso_bridge/striso_util -P | ~/projecten/striso/software/striso_bridge/stribri -p /dev/stdin
+./striso_util -P | ./stribri -p /dev/stdin